Python/Coderforces_Problem_Scrapper/requirements.txt
pillow
fpdf
selenium

